Remove Messaging service from service hub
Remove mention of MessagingService as being on ServiceHub.
