+++ core/src/test/kotlin/net/corda/core/serialization/KryoTests.kt
@@ -3,7 +3,7 @@ package net.corda.core.serialization
 import com.esotericsoftware.kryo.Kryo
 import com.google.common.primitives.Ints
 import net.corda.core.crypto.*
-import net.corda.core.messaging.Ack
+import net.corda.node.services.messaging.Ack
 import net.corda.node.services.persistence.NodeAttachmentService
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
