+++ node/src/integration-test/kotlin/net/corda/services/messaging/P2PMessagingTest.kt
@@ -5,16 +5,16 @@ import com.google.common.util.concurrent.ListenableFuture
 import net.corda.core.*
 import net.corda.core.messaging.MessageRecipients
 import net.corda.core.messaging.SingleMessageRecipient
-import net.corda.core.messaging.createMessage
 import net.corda.core.node.services.DEFAULT_SESSION_ID
 import net.corda.core.node.services.ServiceInfo
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.serialization.deserialize
 import net.corda.core.serialization.serialize
 import net.corda.core.utilities.*
-import net.corda.flows.ServiceRequestMessage
-import net.corda.flows.sendRequest
 import net.corda.node.internal.Node
+import net.corda.node.services.messaging.ServiceRequestMessage
+import net.corda.node.services.messaging.createMessage
+import net.corda.node.services.messaging.sendRequest
 import net.corda.node.services.transactions.RaftValidatingNotaryService
 import net.corda.node.services.transactions.SimpleNotaryService
 import net.corda.node.utilities.ServiceIdentityGenerator
