+++ node/src/main/kotlin/net/corda/node/services/statemachine/StateMachineManager.kt
@@ -18,9 +18,6 @@ import net.corda.core.crypto.Party
 import net.corda.core.crypto.SecureHash
 import net.corda.core.crypto.commonName
 import net.corda.core.flows.*
-import net.corda.core.messaging.ReceivedMessage
-import net.corda.core.messaging.TopicSession
-import net.corda.core.messaging.send
 import net.corda.core.serialization.*
 import net.corda.core.utilities.debug
 import net.corda.core.utilities.loggerFor
@@ -29,6 +26,9 @@ import net.corda.node.internal.ServiceFlowInfo
 import net.corda.node.services.api.Checkpoint
 import net.corda.node.services.api.CheckpointStorage
 import net.corda.node.services.api.ServiceHubInternal
+import net.corda.node.services.messaging.ReceivedMessage
+import net.corda.node.services.messaging.TopicSession
+import net.corda.node.services.messaging.send
 import net.corda.node.utilities.*
 import org.apache.activemq.artemis.utils.ReusableLatch
 import org.jetbrains.exposed.sql.Database
