+++ node/src/test/kotlin/net/corda/node/services/messaging/ArtemisMessagingTests.kt
@@ -5,12 +5,8 @@ import com.google.common.net.HostAndPort
 import com.google.common.util.concurrent.Futures
 import com.google.common.util.concurrent.ListenableFuture
 import com.google.common.util.concurrent.SettableFuture
-import com.typesafe.config.ConfigFactory.empty
-import net.corda.core.crypto.X509Utilities
 import net.corda.core.crypto.generateKeyPair
-import net.corda.core.messaging.Message
 import net.corda.core.messaging.RPCOps
-import net.corda.core.messaging.createMessage
 import net.corda.core.node.services.DEFAULT_SESSION_ID
 import net.corda.core.utilities.ALICE
 import net.corda.core.utilities.LogHelper
