+++ node/src/test/kotlin/net/corda/node/services/network/AbstractNetworkMapServiceTest.kt
@@ -1,16 +1,19 @@
 package net.corda.node.services.network
 
 import com.google.common.util.concurrent.ListenableFuture
-import net.corda.core.crypto.X509Utilities
 import net.corda.core.getOrThrow
 import net.corda.core.messaging.SingleMessageRecipient
-import net.corda.core.messaging.send
 import net.corda.core.node.NodeInfo
 import net.corda.core.node.services.DEFAULT_SESSION_ID
 import net.corda.core.node.services.ServiceInfo
 import net.corda.core.serialization.deserialize
-import net.corda.flows.sendRequest
+import net.corda.core.utilities.ALICE
+import net.corda.core.utilities.BOB
+import net.corda.core.utilities.CHARLIE
+import net.corda.core.utilities.DUMMY_MAP
 import net.corda.node.services.config.NodeConfiguration
+import net.corda.node.services.messaging.send
+import net.corda.node.services.messaging.sendRequest
 import net.corda.node.services.network.AbstractNetworkMapServiceTest.Changed.Added
 import net.corda.node.services.network.AbstractNetworkMapServiceTest.Changed.Removed
 import net.corda.node.services.network.NetworkMapService.*
@@ -20,18 +23,12 @@ import net.corda.node.services.network.NetworkMapService.Companion.PUSH_TOPIC
 import net.corda.node.services.network.NetworkMapService.Companion.QUERY_TOPIC
 import net.corda.node.services.network.NetworkMapService.Companion.REGISTER_TOPIC
 import net.corda.node.services.network.NetworkMapService.Companion.SUBSCRIPTION_TOPIC
-import net.corda.node.services.network.NodeRegistration
-import net.corda.core.utilities.ALICE
-import net.corda.core.utilities.BOB
-import net.corda.core.utilities.CHARLIE
-import net.corda.core.utilities.DUMMY_MAP
 import net.corda.node.utilities.AddOrRemove
 import net.corda.node.utilities.AddOrRemove.ADD
 import net.corda.node.utilities.AddOrRemove.REMOVE
 import net.corda.testing.node.MockNetwork
 import net.corda.testing.node.MockNetwork.MockNode
 import org.assertj.core.api.Assertions.assertThat
-import org.bouncycastle.asn1.x500.X500Name
 import org.eclipse.jetty.util.BlockingArrayQueue
 import org.junit.After
 import org.junit.Before
